Fix typo: Rename 'Extentions' to 'Extensions' everywhere
Corrected namespace and file path from 'Extentions' to 'Extensions' in code, using statements, and documentation. This ensures consistency and prevents reference errors across the project.
